requirements.txt
boto3
moviepy
python-dotenv
tweepy
kaggle